move server back
